# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TFG API Segura - Gestión de Informes Médicos",
        "description": "API REST segura con autenticación JWT, MFA, roles y permisos, rate limiting y auditoría de seguridad.",
        "contact": {
            "name": "Pablo Sánchez",
            "email": "pablo@tfg.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/mfa/setup": {
            "post": {
                "tags": [
                    "MFA"
                ],
                "summary": "Iniciar configuración de MFA",
                "operationId": "26847e5b34e6efe42170c4bcbcb99912",
                "responses": {
                    "200": {
                        "description": "Secreto y QR generados"
                    },
                    "400": {
                        "description": "MFA ya está activado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mfa/confirm": {
            "post": {
                "tags": [
                    "MFA"
                ],
                "summary": "Confirmar y activar MFA",
                "operationId": "c66baced57c785170f687f32bbdc95a4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "MFA activado correctamente"
                    },
                    "422": {
                        "description": "Código inválido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mfa/disable": {
            "post": {
                "tags": [
                    "MFA"
                ],
                "summary": "Desactivar MFA",
                "operationId": "9a69a35b6eeb5459cc7fca8eaad7df9e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "MFA desactivado correctamente"
                    },
                    "422": {
                        "description": "Código inválido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mfa/verify-login": {
            "post": {
                "tags": [
                    "MFA"
                ],
                "summary": "Verificar código MFA durante el login",
                "operationId": "1e11f0b6ea4c49869a93c4f149a6bc76",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "temporary_token",
                                    "code"
                                ],
                                "properties": {
                                    "temporary_token": {
                                        "type": "string",
                                        "example": "abc123..."
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login completado con éxito"
                    },
                    "401": {
                        "description": "Token o código inválido"
                    }
                }
            }
        },
        "/mfa/recovery-codes/regenerate": {
            "post": {
                "tags": [
                    "MFA"
                ],
                "summary": "Regenerar códigos de recuperación MFA",
                "operationId": "f60e1b846a8f0905e1adc7523763d791",
                "responses": {
                    "200": {
                        "description": "Códigos regenerados"
                    },
                    "400": {
                        "description": "MFA no está activado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}